+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Version 1.0.1 (2012-12-14)
 ### Fixed
 - GET params were transformed on __get() instead of __set()
 - Action form parameter was not correctly set
+- Made the fields tableless by default
 
 Version 1.0.0 (2012-10-05)
 ------------------------------